Do not accept the same uid multiple times; specify ByteCounter.
This requires that the callbacks are informed of the UID so that it can be
removed from the set of accepted probes upon completion. The implemention
of this is incredibly messy - it could be improved by wrapping the
public static variables in methods to guard their usage, but hopefully
there's a fundamentally different approach to use.

ByteCounter was null, which prevents messages from being sent.