Remove overly verbose logging.
Acceptance is logged anyway and then only if it has not already been
accepted, which is more useful.

Remove logging callback on sending relay as it's established that earlier
problems weren't due to the message not being received.

Result callback logging is kinda silly because the purpose of the callback
itself is to inform of these events anyway.

Also fixes some whitespace in a logging statement.