Check for free slot first; reserve slot last.
This is so messages are dropped due to lack of slots as soon as
possible. However, this means that if multiple requests are
received in quick succession, they could each pass the check if not
enough reach the reservation code, and the node could commit to more
than the limit.