Check for a free slot right after reading UID.
This is as soon as I can think to make it, and leaves reserving a new slot
at the end where it means something wrong with the rest doesn't require
removing it again before returning. However, this means that if multiple
requests are received in quick succession, they could each pass the check
if not enough reach the reservation code, and the node could commit to more
than the limit.